Refactored frontend structure
