+++ frontend/login.html
@@ -7,7 +7,7 @@
     <link rel="stylesheet" href="./src/css/common.css">
     <link rel="stylesheet" href="./src/css/login.css">
     <link rel="icon" href="NP_Simple_Circle.png" type="image/png">
-    <script src="./src/login.ts" type="module"></script>
+    <script src="./src/pages/login.ts" type="module"></script>
 </head>
 
 <body>
