+++ frontend/main.html
@@ -7,7 +7,7 @@
     <link rel="stylesheet" href="./src/css/common.css">
     <link rel="stylesheet" href="./src/css/main.css">
     <link rel="icon" href="NP_Simple_Circle.png" type="image/png">
-    <script src="./src/main.ts" type="module"></script>
+    <script src="./src/pages/main.ts" type="module"></script>
 </head>
 
 <body>
