+++ frontend/src/lib/enums.ts
