+++ frontend/src/lib/pending_updates.ts
