+++ frontend/src/windows/map_list_window.ts
@@ -1,10 +1,10 @@
-import * as ContextMenu from "./contextmenu.ts";
-import { ApiRequest, Session } from "./requests.ts";
+import * as ContextMenu from "../lib/contextmenu.ts";
+import { ApiRequest, Session } from "../lib/requests.ts";
 import { ContentWindow, InputDialog, registerWindowType } from "./window.ts";
-import { Parameter, AddDragListener } from "./utils.ts";
-import { Vector2 } from "./vector.ts";
-import { ErrorToast } from "./notifications.ts";
-import { Html } from "./elements.ts";
+import { Parameter, AddDragListener } from "../lib/utils.ts";
+import { Vector2 } from "../vector.ts";
+import { ErrorToast } from "../lib/notifications.ts";
+import { Html } from "../lib/elements.ts";
 import { MapWindow } from "./map_window.ts";
